refactor(providers): share FAT16 image encoding
